Move retriever files into a separate directory
