+++ genai-perf/genai_perf/inputs/retrievers/synthetic_image_generator.py
